core.plot.skparplot (and probably dftbutils.plot.magic_bs_plot) are made to run via a wrapper callable (essentially the old PlotTask). All tests now pass, but the plot functionality is not tested via yaml input
